add type badge
